// src/interfaces.ts
export interface ICharacter {
  birth_year?: string;
  eye_color?: string;
  films?: string[];
  gender?: any;
  hair_color?: string;
  height?: string;
  homeworld?: any;
  mass?: string;
  name?: any;
  skin_color?: string;
  created?: string;
  edited?: string;
  species?: string[];
  starships?: string[];
  url?: string;
  vehicles?: string[];
  favorites?: boolean;
  moreInfo?: boolean;
}